Initial Commit
This is the introduction of Hussh.
I'm hopeful that this project will be an easy way for any Python
developer to get very fast ssh interactions with no real effort on their
part.
I have a lot to learn to get this project where I want it, so let's get
started!