Add explicit close and context manager to Connection
This commit introduces a new close method that will close the session on
a Connection object. However, ssh2-rs states that this will not close
the underlying socket. If that's an issue in the future, perhaps we also
track the TCP connection and close that along with the session.

Since we're now allowing explicit closing, I also added a context
manager to the Connection class, so people can handle that implicitly.

I also bumped the supported Python versions included 3.13 (and
pre-releases).

Fixes #15